Renaming and update to postgresql 9.6
There were several uncertainties with used names. To clarify them a
big rework was done. Furthermore there were added two tables so that
the structure of used jenkins jobs maps to the structure of the
database. An additional table stores the min/avg/max values of the
cyclictest runs per thread, so that there is no missing cyclictest
information.

Debain Stretch uses postgresql 9.6 instead of 9.5 and 9.6 is the LTS
version. Therefore update the scheme to 9.6 version.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>