Change PostgreSQL database schema from public to ci-rt
The "public" schema is open to all users by default.
>From the PostgreSQL Manual:

5.8.4. Schemas and Privileges

A user can also be allowed to create objects in someone else's schema. To
allow that, the CREATE privilege on the schema needs to be granted. Note
that by default, everyone has CREATE and USAGE privileges on the schema
public. This allows all users that are able to connect to a given database
to create objects in its public schema.

Use a user defined schema "ci-rt" to handle database permissions in a
compatible way.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>