Rework database permissions
When all users are allowed to use and change database schema 'ci-rt', a
missconfigured database could be the result. Therefore do not allow users
of database schema "ci-rt" to create tables. And do not allow other users
than jenkins and consumer to use the schema.

While at it update description for database schema 'ci-rt'.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>