Remove quotation marks from comments
While changing the database schema from "public" to "ci-rt" the new
schema is referenced in the comments with "ci-rt".
A dump of the database structure with pg_dump references the same
schema as ci-rt. Change the comments to be able to diff the output
of pg_dump with the database structure without annoying noise.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>